api/requirements.txt
crewai==0.11.0
anthropic==0.7.8
openai==1.12.0
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.0
motor==3.3.2
stability-sdk==0.8.3
pydantic==2.6.1
tenacity==8.2.3
python-multipart==0.0.9
pillow==10.2.0
redis==5.0.1